refactor move enum to data
