+++ src/main/java/mrriegel/storagenetwork/block/request/TileRequest.java
@@ -3,6 +3,7 @@ package mrriegel.storagenetwork.block.request;
 import java.util.HashMap;
 import java.util.Map;
 import mrriegel.storagenetwork.block.TileConnectable;
+import mrriegel.storagenetwork.util.data.EnumSortType;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.nbt.NBTTagList;
