+++ src/main/java/mrriegel/storagenetwork/item/remote/ItemRemote.java
@@ -5,10 +5,10 @@ import javax.annotation.Nullable;
 import mrriegel.storagenetwork.CreativeTab;
 import mrriegel.storagenetwork.StorageNetwork;
 import mrriegel.storagenetwork.block.master.TileMaster;
-import mrriegel.storagenetwork.block.request.EnumSortType;
 import mrriegel.storagenetwork.config.ConfigHandler;
 import mrriegel.storagenetwork.gui.GuiHandler;
 import mrriegel.storagenetwork.util.NBTHelper;
+import mrriegel.storagenetwork.util.data.EnumSortType;
 import net.minecraft.client.resources.I18n;
 import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.creativetab.CreativeTabs;
