+++ src/main/java/mrriegel/storagenetwork/util/data/EnumSortType.java
@@ -1,4 +1,4 @@
-package mrriegel.storagenetwork.block.request;
+package mrriegel.storagenetwork.util.data;
 
 public enum EnumSortType {
   AMOUNT, NAME, MOD;
